refactors common code to project
