+++ Mindscape.Raygun4Net.AspNetCore.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mindscape.Raygun4Net.NetCor
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mindscape.Raygun4Net.AspNetCore2.Tests", "Mindscape.Raygun4Net.AspNetCore2.Tests\Mindscape.Raygun4Net.AspNetCore2.Tests.csproj", "{1B8F6225-7D9C-4D4A-A2AD-8F78F8FA8B8B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mindscape.Raygun4Net.NetCore.Common", "Mindscape.Raygun4Net.NetCore.Common\Mindscape.Raygun4Net.NetCore.Common.csproj", "{A069590F-1E39-4EFF-AEB3-437E4538DA8E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -55,6 +57,18 @@ Global
 		{1B8F6225-7D9C-4D4A-A2AD-8F78F8FA8B8B}.Release|x64.Build.0 = Release|Any CPU
 		{1B8F6225-7D9C-4D4A-A2AD-8F78F8FA8B8B}.Release|x86.ActiveCfg = Release|Any CPU
 		{1B8F6225-7D9C-4D4A-A2AD-8F78F8FA8B8B}.Release|x86.Build.0 = Release|Any CPU
+		{A069590F-1E39-4EFF-AEB3-437E4538DA8E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A069590F-1E39-4EFF-AEB3-437E4538DA8E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A069590F-1E39-4EFF-AEB3-437E4538DA8E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A069590F-1E39-4EFF-AEB3-437E4538DA8E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A069590F-1E39-4EFF-AEB3-437E4538DA8E}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{A069590F-1E39-4EFF-AEB3-437E4538DA8E}.Debug|x64.Build.0 = Debug|Any CPU
+		{A069590F-1E39-4EFF-AEB3-437E4538DA8E}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{A069590F-1E39-4EFF-AEB3-437E4538DA8E}.Debug|x86.Build.0 = Debug|Any CPU
+		{A069590F-1E39-4EFF-AEB3-437E4538DA8E}.Release|x64.ActiveCfg = Release|Any CPU
+		{A069590F-1E39-4EFF-AEB3-437E4538DA8E}.Release|x64.Build.0 = Release|Any CPU
+		{A069590F-1E39-4EFF-AEB3-437E4538DA8E}.Release|x86.ActiveCfg = Release|Any CPU
+		{A069590F-1E39-4EFF-AEB3-437E4538DA8E}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
