+++ Mindscape.Raygun4Net.NetCore.Common/Builders/RaygunErrorMessageBuilderBase.cs
@@ -1,0 +1,191 @@
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Reflection;
+using System.Text;
+
+namespace Mindscape.Raygun4Net
+{
+  public abstract class RaygunErrorMessageBuilderBase
+  {
+    protected static string FormatTypeName(Type type, bool fullName)
+    {
+      string name = fullName ? type.FullName : type.Name;
+      if (!type.IsConstructedGenericType)
+      {
+        return name;
+      }
+
+      StringBuilder stringBuilder = new StringBuilder();
+      stringBuilder.Append(name.Substring(0, name.IndexOf("`")));
+      stringBuilder.Append("<");
+      foreach (Type t in type.GenericTypeArguments)
+      {
+        stringBuilder.Append(FormatTypeName(t, false)).Append(",");
+      }
+      stringBuilder.Remove(stringBuilder.Length - 1, 1);
+      stringBuilder.Append(">");
+
+      return stringBuilder.ToString();
+    }
+    
+    protected static RaygunErrorStackTraceLineMessage[] BuildStackTrace(Exception exception)
+    {
+      var lines = new List<RaygunErrorStackTraceLineMessage>();
+
+      var stackTrace = new StackTrace(exception, true);
+      var frames = stackTrace.GetFrames();
+
+      if (frames == null || frames.Length == 0)
+      {
+        var line = new RaygunErrorStackTraceLineMessage { FileName = "none", LineNumber = 0 };
+        lines.Add(line);
+        
+        return lines.ToArray();
+      }
+
+      foreach (StackFrame frame in frames)
+      {
+        MethodBase method = frame.GetMethod();
+
+        if (method != null)
+        {
+          int lineNumber = frame.GetFileLineNumber();
+
+          if (lineNumber == 0)
+          {
+            lineNumber = frame.GetILOffset();
+          }
+
+          var methodName = GenerateMethodName(method);
+
+          string file = frame.GetFileName();
+
+          string className = method.DeclaringType != null ? method.DeclaringType.FullName : "(unknown)";
+
+          var line = new RaygunErrorStackTraceLineMessage
+          {
+            FileName = file,
+            LineNumber = lineNumber,
+            MethodName = methodName,
+            ClassName = className
+          };
+
+          lines.Add(line);
+        }
+      }
+
+      return lines.ToArray();
+    }
+    
+    protected static string GenerateMethodName(MethodBase method)
+    {
+      var stringBuilder = new StringBuilder();
+
+      stringBuilder.Append(method.Name);
+
+      bool first = true;
+      
+      if (method is MethodInfo && method.IsGenericMethod)
+      {
+        Type[] genericArguments = method.GetGenericArguments();
+        stringBuilder.Append("[");
+        
+        for (int i = 0; i < genericArguments.Length; i++)
+        {
+          if (!first)
+          {
+            stringBuilder.Append(",");
+          }
+          else
+          {
+            first = false;
+          }
+          
+          stringBuilder.Append(genericArguments[i].Name);
+        }
+        
+        stringBuilder.Append("]");
+      }
+      
+      stringBuilder.Append("(");
+      
+      ParameterInfo[] parameters = method.GetParameters();
+      
+      first = true;
+      
+      for (int i = 0; i < parameters.Length; ++i)
+      {
+        if (!first)
+        {
+          stringBuilder.Append(", ");
+        }
+        else
+        {
+          first = false;
+        }
+        
+        string type = "<UnknownType>";
+        
+        if (parameters[i].ParameterType != null)
+        {
+          type = parameters[i].ParameterType.Name;
+        }
+        
+        stringBuilder.Append(type + " " + parameters[i].Name);
+      }
+      
+      stringBuilder.Append(")");
+
+      return stringBuilder.ToString();
+    }
+    
+    public static RaygunErrorMessage Build(Exception exception)
+    {
+      RaygunErrorMessage message = new RaygunErrorMessage();
+
+      var exceptionType = exception.GetType();
+
+      message.Message = exception.Message;
+      message.ClassName = FormatTypeName(exceptionType, true);
+
+      message.StackTrace = BuildStackTrace(exception);
+
+      if (exception.Data != null)
+      {
+        IDictionary data = new Dictionary<object, object>();
+        
+        foreach (object key in exception.Data.Keys)
+        {
+          if (!RaygunClientBase.SentKey.Equals(key))
+          {
+            data[key] = exception.Data[key];
+          }
+        }
+        
+        message.Data = data;
+      }
+
+      AggregateException ae = exception as AggregateException;
+      
+      if (ae != null && ae.InnerExceptions != null)
+      {
+        message.InnerErrors = new RaygunErrorMessage[ae.InnerExceptions.Count];
+        int index = 0;
+        
+        foreach (Exception e in ae.InnerExceptions)
+        {
+          message.InnerErrors[index] = Build(e);
+          index++;
+        }
+      }
+      else if (exception.InnerException != null)
+      {
+        message.InnerError = Build(exception.InnerException);
+      }
+
+      return message;
+    }
+  }
+}
