+++ Mindscape.Raygun4Net.NetCore.Common/Messages/RaygunClientMessage.cs
