+++ Mindscape.Raygun4Net.NetCore.Common/Messages/RaygunEnvironmentMessage.cs
