+++ Mindscape.Raygun4Net.NetCore.Common/Messages/RaygunErrorMessage.cs
