+++ Mindscape.Raygun4Net.NetCore.Common/Messages/RaygunErrorStackTraceLineMessage.cs
