+++ Mindscape.Raygun4Net.NetCore.Common/Messages/RaygunIdentifierMessage.cs
