+++ Mindscape.Raygun4Net.NetCore.Common/Messages/RaygunMessage.cs
