+++ Mindscape.Raygun4Net.NetCore.Common/Messages/RaygunMessageDetails.cs
