+++ Mindscape.Raygun4Net.NetCore.Common/Messages/RaygunRequestMessage.cs
