+++ Mindscape.Raygun4Net.NetCore.Common/Messages/RaygunResponseMessage.cs
