+++ Mindscape.Raygun4Net.NetCore.Common/RaygunCustomGroupingKeyEventArgs.cs
