+++ Mindscape.Raygun4Net.NetCore.Common/RaygunSendingMessageEventArgs.cs
