+++ Mindscape.Raygun4Net.NetCore.Common/SimpleJson.cs
