--- Mindscape.Raygun4Net.NetCore/Builders/RaygunErrorMessageBuilder.cs
@@ -1,107 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Reflection;
-
-namespace Mindscape.Raygun4Net
-{
-  public class RaygunErrorMessageBuilder : RaygunErrorMessageBuilderBase
-  {
-    public static RaygunErrorMessage Build(Exception exception)
-    {
-      RaygunErrorMessage message = new RaygunErrorMessage();
-
-      var exceptionType = exception.GetType();
-
-      message.Message = exception.Message;
-      message.ClassName = FormatTypeName(exceptionType, true);
-
-      message.StackTrace = BuildStackTrace(exception);
-
-      if (exception.Data != null)
-      {
-        IDictionary data = new Dictionary<object, object>();
-        
-        foreach (object key in exception.Data.Keys)
-        {
-          if (!RaygunClient.SentKey.Equals(key))
-          {
-            data[key] = exception.Data[key];
-          }
-        }
-        
-        message.Data = data;
-      }
-
-      AggregateException ae = exception as AggregateException;
-      
-      if (ae != null && ae.InnerExceptions != null)
-      {
-        message.InnerErrors = new RaygunErrorMessage[ae.InnerExceptions.Count];
-        int index = 0;
-        
-        foreach (Exception e in ae.InnerExceptions)
-        {
-          message.InnerErrors[index] = Build(e);
-          index++;
-        }
-      }
-      else if (exception.InnerException != null)
-      {
-        message.InnerError = Build(exception.InnerException);
-      }
-
-      return message;
-    }
-
-    private static RaygunErrorStackTraceLineMessage[] BuildStackTrace(Exception exception)
-    {
-      var lines = new List<RaygunErrorStackTraceLineMessage>();
-
-      var stackTrace = new StackTrace(exception, true);
-      var frames = stackTrace.GetFrames();
-
-      if (frames == null || frames.Length == 0)
-      {
-        var line = new RaygunErrorStackTraceLineMessage { FileName = "none", LineNumber = 0 };
-        lines.Add(line);
-        
-        return lines.ToArray();
-      }
-
-      foreach (StackFrame frame in frames)
-      {
-        MethodBase method = frame.GetMethod();
-
-        if (method != null)
-        {
-          int lineNumber = frame.GetFileLineNumber();
-
-          if (lineNumber == 0)
-          {
-            lineNumber = frame.GetILOffset();
-          }
-
-          var methodName = GenerateMethodName(method);
-
-          string file = frame.GetFileName();
-
-          string className = method.DeclaringType != null ? method.DeclaringType.FullName : "(unknown)";
-
-          var line = new RaygunErrorStackTraceLineMessage
-          {
-            FileName = file,
-            LineNumber = lineNumber,
-            MethodName = methodName,
-            ClassName = className
-          };
-
-          lines.Add(line);
-        }
-      }
-
-      return lines.ToArray();
-    }
-  }
-}
